requirements.txt
numpy
fpdf
pillow
matplotlib
flask
flask_cors
requests
pandas
